requirements.txt
annotated-types==0.7.0
anyio==4.8.0
dnspython==1.16.0
exceptiongroup==1.2.2
fastapi==0.115.7
idna==3.10
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.11
sniffio==1.3.1
starlette==0.45.2
typing_extensions==4.12.2
motor==3.7.0
firebase-admin==6.6.0
Pyrebase4==4.8.0
setuptools==75.8.0
python-dotenv==1.0.1
pyjwt==2.10.0
passlib==1.7.4